+++ packages/components/src/components/hds/pagination/numbered/index.ts
@@ -6,7 +6,6 @@ import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { assert } from '@ember/debug';
-import { elliptize } from '../elliptize.ts';
 import { HdsPaginationDirectionValues } from '../types.ts';
 
 import type {
@@ -16,6 +15,11 @@ import type {
   HdsPaginationElliptizedPageArrayItem,
 } from '../types';
 import type { HdsInteractiveSignature } from '../../interactive/index.ts';
+interface ElliptizeProps {
+  pages: number[];
+  current: number;
+  limit?: number;
+}
 
 type HdsInteractiveQuery = HdsInteractiveSignature['Args']['query'];
 
@@ -51,9 +55,71 @@ export interface HdsPaginationNumberedSignature {
   Element: HTMLDivElement;
 }
 
+const ELLIPSIS = '…';
+
 // for context about the decision to use these values, see:
 // https://hashicorp.slack.com/archives/C03A0N1QK8S/p1673546329082759
 export const DEFAULT_PAGE_SIZES = [10, 30, 50];
+
+const elliptize = ({
+  pages,
+  current,
+  limit = 7,
+}: ElliptizeProps): HdsPaginationElliptizedPageArray => {
+  const length = pages.length;
+
+  let result = [];
+  let start;
+  let end;
+
+  if (length <= limit) {
+    return pages;
+  }
+
+  if (current <= length / 2) {
+    start = Math.ceil(limit / 2);
+    end = limit - start;
+  } else {
+    end = Math.ceil(limit / 2);
+    start = limit - end;
+  }
+
+  const sliceStart: HdsPaginationElliptizedPageArray = pages.slice(0, start);
+  const sliceEnd: HdsPaginationElliptizedPageArray = pages.slice(-end);
+
+  if (sliceStart.includes(current) && sliceStart.includes(current + 1)) {
+    // "current" (and its next sibling) is contained within the "sliceStart" block
+    sliceEnd.splice(0, 1, ELLIPSIS);
+    result = ([] as HdsPaginationElliptizedPageArray).concat(
+      sliceStart,
+      sliceEnd
+    );
+  } else if (sliceEnd.includes(current - 1) && sliceEnd.includes(current)) {
+    // "current" (and its prev sibling) is contained within the "sliceEnd" block
+    sliceStart.splice(-1, 1, ELLIPSIS);
+    result = ([] as HdsPaginationElliptizedPageArray).concat(
+      sliceStart,
+      sliceEnd
+    );
+  } else {
+    // this is a bit more tricky :)
+    // we need to calculate how many items there are before/after the current item
+    // since both the initial and ending blocks are always 2 items long (number + ellipsis)
+    // and there is always the "current" item, we can just subtract 5 from the limit
+    const delta = (limit - 5) / 2; // this is why the limit needs to be an odd number
+    // we slice the array starting at the "current" index, minus the delta, minus one because it's an array (zero-based)
+    const sliceCurr = pages.slice(current - delta - 1, current + delta);
+    result = ([] as HdsPaginationElliptizedPageArray).concat(
+      sliceStart.shift() as number,
+      ELLIPSIS,
+      sliceCurr,
+      ELLIPSIS,
+      sliceEnd.pop() as number
+    );
+  }
+
+  return result;
+};
 export default class HdsPaginationNumberedComponent extends Component<HdsPaginationNumberedSignature> {
   // These two private variables are used to differentiate between
   // "uncontrolled" component (where the state is handled internally) and
